app/src/main/java/com/example/appstory/view/listStory/ListStoryViewModel.kt
package com.example.appstory.view.listStory

import androidx.lifecycle.LiveData
import androidx.lifecycle.MutableLiveData
import androidx.lifecycle.ViewModel
import androidx.lifecycle.viewModelScope
import androidx.paging.PagingData
import androidx.paging.cachedIn
import com.example.appstory.Data.Retrofit.response.ListStoryItem
import com.example.appstory.Data.repository.StoryRepository
import kotlinx.coroutines.flow.MutableStateFlow
import kotlinx.coroutines.flow.asStateFlow
import kotlinx.coroutines.flow.update
import kotlinx.coroutines.launch


class ListStoryViewModel(private val storyRepository: StoryRepository): ViewModel() {

    val quote: LiveData<PagingData<ListStoryItem>> =
        storyRepository.getStory().cachedIn(viewModelScope)
}